requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM
openai==1.58.1

# HTTP client
httpx==0.28.1

# CORS
# (included in FastAPI)

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
authlib==1.3.2
itsdangerous==2.2.0

# Email
fastapi-mail==1.4.1

# Dev
black==24.10.0
ruff==0.8.6
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
